fix sys_ioctl.c

+++ library/sys_ioctl.c
@@ -1,7 +1,7 @@
 #include <sys/ioctl.h>
 
 int __stdargs
-ioctl (int fd, unsigned long cmd, ...)
+ioctl (int fd, int cmd, ...)
 {
   return -1;
 }
